Refactor PDF generation into modular package structure
- Break down 522-line monolithic generate_pdf.py into focused modules:
  - pdf_generator/css_styles.py - CSS styling (198 lines)
  - pdf_generator/markdown_processor.py - Markdown processing (72 lines)
  - pdf_generator/file_utils.py - File discovery utilities (32 lines)
  - pdf_generator/pdf_builder.py - PDF generation logic (148 lines)
  - generate_pdfs.py - Main orchestrator (54 lines, 90% reduction)
- Create proper Python package with __init__.py for clean imports
- Update generate_pdfs.sh to call new script name
- Fix markdown newline processing that was showing literal '\n\n' in PDFs
- Maintain full functionality while dramatically improving maintainability
- Each module now has single responsibility and can be tested independently
